feat: reorganize acme-core into config/provider subpackages Mirrors acme-spring's structure: config-contract classes in .config, acme4j SPI plumbing in .provider, the actual client API (AcmeService and friends) stays at the top level. Refs: XRDDEV-3286
