+++ src/lib/acme-core/src/main/java/org/niis/xroad/common/acme/AcmeService.java
@@ -33,6 +33,9 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.bouncycastle.operator.OperatorCreationException;
 import org.jose4j.jws.AlgorithmIdentifiers;
+import org.niis.xroad.common.acme.config.AcmeConfig;
+import org.niis.xroad.common.acme.config.AcmeProperties;
+import org.niis.xroad.common.acme.provider.AcmeCustomSchema;
 import org.niis.xroad.common.exception.BadRequestException;
 import org.niis.xroad.globalconf.model.ApprovedCAInfo;
 import org.shredzone.acme4j.Account;
