+++ src/lib/acme-core/src/main/resources/META-INF/services/org.shredzone.acme4j.provider.AcmeProvider
@@ -1,1 +1,1 @@
-org.niis.xroad.common.acme.AcmeXroadProvider
+org.niis.xroad.common.acme.provider.AcmeXroadProvider
