+++ src/lib/acme-core/src/test/java/org/niis/xroad/common/acme/AcmeServiceTest.java
@@ -38,6 +38,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 import org.junit.jupiter.api.io.TempDir;
+import org.niis.xroad.common.acme.config.AcmeConfig;
+import org.niis.xroad.common.acme.config.AcmeProperties;
 import org.niis.xroad.common.exception.NotFoundException;
 import org.niis.xroad.globalconf.model.ApprovedCAInfo;
