+++ src/lib/acme-spring/src/main/java/org/niis/xroad/common/acme/spring/config/AcmeChallengeFilter.java
@@ -34,7 +34,7 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.niis.xroad.common.acme.AcmeConfig;
+import org.niis.xroad.common.acme.config.AcmeConfig;
 import org.springframework.stereotype.Component;
 
 import java.io.IOException;
