+++ src/security-server/admin-service/application/src/main/java/org/niis/xroad/securityserver/restapi/scheduling/AcmeClientWorker.java
@@ -34,9 +34,9 @@ import ee.ria.xroad.common.identifier.SecurityServerId;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.niis.xroad.common.acme.AcmeConfig;
 import org.niis.xroad.common.acme.AcmeKeyPurpose;
 import org.niis.xroad.common.acme.AcmeService;
+import org.niis.xroad.common.acme.config.AcmeConfig;
 import org.niis.xroad.common.acme.spring.scheduling.AcmeRenewalWorker;
 import org.niis.xroad.common.acme.spring.scheduling.CertificateRenewalScheduler;
 import org.niis.xroad.common.core.exception.XrdRuntimeException;
